% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{palette_bold}
\alias{palette_bold}
\title{boldR colour palette}
\format{
A named character vector of length 8.
}
\usage{
palette_bold
}
\description{
A named character vector of the eight colours used throughout the boldR
package and its pkgdown site.
}
